test: move Rhythm tests to their own descriptors/rhythm.cpp
Follow the one-cpp-per-descriptor layout requested in review (#118): put
the Rhythm test cases in tests/descriptors/rhythm.cpp and register it in the
test_descriptors executable, instead of appending to the monolithic
test_descriptors.cpp. No test logic changed.
